requirements.txt
# Core
tensorflow>=2.15.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
Pillow>=10.0.0
matplotlib>=3.8.0

# App
streamlit>=1.36.0

# Utils
tqdm>=4.66.0
